Make DashBoardAccordion a separate component
