+++ frontend/main-app/src/components/DashboardAccordion/DashboardAccordion.module.css
@@ -1,0 +1,47 @@
+.formList {
+    display: flex;
+    flex-direction: column;
+    gap: var(--fds-spacing-2);
+}
+
+.accordionContentContainer {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: var(--fds-spacing-4);
+}
+
+.buttonContainer {
+    padding-top: var(--fds-spacing-1);
+    display: flex;
+    flex-direction: column;
+    gap: var(--fds-spacing-4);
+}
+
+.button {
+    justify-content: left;
+}
+
+@media screen and (max-width: 1000px) {
+    .accordionContentContainer {
+        flex-direction: column;
+    }
+
+    .infoContainer {
+        margin-bottom: var(--fds-spacing-4);
+    }
+
+    .buttonContainer {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+    }
+
+    .button {
+        justify-content: center;
+    }
+}
+
+@media screen and (max-width: 550px) {
+    .buttonContainer {
+        grid-template-columns: 1fr;
+    }
+}
